requirements.txt
annotated-types==0.7.0
anyio==4.13.0
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.1.8
colorama==0.4.6
cryptography==46.0.7
distro==1.9.0
duckduckgo_search==8.1.1
filelock==3.28.0
fsspec==2026.3.0
google-auth==2.49.2
google-genai==1.73.1
greenlet==3.4.0
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.14.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.1.1
langchain==0.3.25
langchain-core==0.3.84
langchain-text-splitters==0.3.11
langsmith==0.3.45
llvmlite==0.47.0
lxml==6.0.4
MarkupSafe==3.0.3
more-itertools==11.0.2
mpmath==1.3.0
networkx==3.6.1
numba==0.65.0
numpy==2.4.4
openai==2.32.0
openai-whisper==20250625
orjson==3.11.8
ormsgpack==1.12.2
packaging==25.0
pillow==12.2.0
primp==1.2.2
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.2
pydantic_core==2.46.2
pygame==2.6.1
python-dotenv==1.2.2
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
requests-toolbelt==1.0.0
scikit-learn==1.8.0
scipy==1.17.1
sentencepiece==0.2.1
setuptools==81.0.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.5
soupsieve==2.8.3
SQLAlchemy==2.0.49
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tiktoken==0.12.0
torch==2.11.0
torchaudio==2.11.0
torchtext==0.6.0
torchvision==0.26.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
websockets==16.0
whisper==1.1.10
xxhash==3.6.0
zstandard==0.23.0
